[DDW-863]: Update transaction components to make reusable component
