+++ source/features/transactions/components/TransactionList.scss
@@ -1,7 +1,4 @@
 .transactionListContainer {
-  margin: 61.5px 0;
-  width: 872px;
-
   .header {
     margin-bottom: 41px;
   }
