+++ source/features/transactions/components/TransactionList.tsx
@@ -1,12 +1,10 @@
 import { observer } from 'mobx-react-lite';
 import DividerWithTitle from '../../widgets/divider-with-title/components/DividerWithTitle';
 import styles from './TransactionList.scss';
-import TransactionListRow, {
-  ITransactionListRowProps,
-} from './TransactionListRow';
+import TransactionInfo, { ITransactionInfoProps } from './TransactionInfo';
 
 export interface ITransactionListProps {
-  items: Array<ITransactionListRowProps>;
+  items: Array<ITransactionInfoProps>;
   title: string;
 }
 
@@ -17,7 +15,7 @@ const TransactionList = (props: ITransactionListProps) => (
     </div>
     {props.items.map((item, index) => (
       <div key={`transaction_${index}`} className={styles.transactionListRow}>
-        <TransactionListRow {...item} />
+        <TransactionInfo {...item} />
       </div>
     ))}
   </div>
